\name{acceptance}
\alias{acceptance}
\title{List the History of Acceptance.}
\description{
 This function reads \file{accept.xml} in \code{.Library}, if it exists, and prints it
 with the default method. 
}
\usage{
acceptance()
}
\details{
  Data is displayed as XML markup.
}
\value{
 an object of class \code{XMLNode}
}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item \code{\link{accept}}
		\item \code{\link{packageCheck}}
	}
}
\keyword{manip}

