\name{backtrans}
\alias{backtrans}

\title{Back-transform Columns in a Data Frame}
\description{
 Exponentiate presumably log-transformed columns in a data.frame.
}
\usage{
backtrans(x, cols)
}

\arguments{
  \item{x}{data.frame}
  \item{cols}{vector of column names}
}
\value{data.frame}
\references{mifuns.google.code.com}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item \code{\link{PLOTR}}
	}
}
\keyword{manip}

