\name{ctl}
\alias{ctl}
\docType{data}
\title{A Sample NONMEM(r) Control Stream}
\description{
  The control stream for running control test 8 of NMQual from Metrum Institute.
}
\usage{data(ctl)}
\format{
  The format is:
 chr [1:89] "PROBLEM 8, Parent-Metabolite-Urine Model" ...
}
\details{
  This control stream is stored here as a practice target for 
  \code{resample.character}.
}
\source{
  The file was borrowed from 8.ctl as distributed with NMQual from
  http://nmqual.googlecode.com.
}
\examples{
data(ctl)
}
\keyword{datasets}

