\name{getPars}
\alias{getPars}

\title{Return the Non-duplicated (ID) Rows Corresponding to the Specified Filename}
\description{Useful for retrieving parameters that are constant within subject.}
\usage{
getPars(file)
}
\arguments{
	\item{file}{path for a file, typcally \file{*par.TAB}}
}
\value{data.frame}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item \code{\link{PLOTR}}
	}
}
\keyword{manip}

