\name{helpAdminister}
\alias{helpAdminister}
\docType{package}
\title{MIfuns: Administrative}
\description{Tools for administrative tasks.}

\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=CLNR]{delete} files and directories; log what was deleted to a text file
\item \link[=hash]{embed} R output in a script
\item \link[=accept]{document} acceptance of an R installation
\item \link[=isSubversioned]{check} wether your files are subversioned
\item \link[=svnIsText]{test} whether Subversion thinks your files are text
\item \link[=svnMarkAsText]{tell} Subversion to treat your files as text
\item \link[=electronicAppendix]{create} an FDA-style electronic appendix from a Subversioned directory
\item \link[=variants]{locate} path-variant versions of a file
 }
}

