\name{helpDataFrame}
\alias{helpDataFrame}
\docType{package}
\title{metrumrg: Working with Data Frames}
\description{Tools for processing data.frames.}
\details{
\link[=metrumrg]{metrumrg} > \link[=helpPrepare]{Data}

I want to ....
\itemize{
\item \link[=shuffle]{move columns in a data frame}
\item \link[=colname<-]{rename selected columns}
\item \link[=fixedwidth]{format a data.frame as fixed-width}
\item \link[=aug]{add named elements to a data.frame or list}
\item \link[=static]{find the static portion of a data frame}
\item \link[=digest]{diagnose and visualize static data relationships}
\item \link[=nasum]{diagnose missingness}
\item \link[=combinations]{show sorted unique rows}
\item \link[=supp]{add supplemental data to a CDISC data set}
\item \link[=specification]{document and enforce data set properties}
}
}
