\name{helpMatrix}
\alias{helpMatrix}
\docType{package}
\title{metrumrg: Working with Matrices}
\description{Tools for processing matrices.}
\details{
\link[=metrumrg]{metrumrg} > \link[=helpPrepare]{Data}

I want to ....
\itemize{
\item \link[=half.matrix]{interconvert} between square matrices and triangular matrices 
}
}
