\name{helpPharmacometric}
\alias{helpPharmacometric}
\docType{package}
\title{MIfuns: Pharmacometrics}
\description{Tools for investigating pharmacometric properties.}
\details{
\link[=MIfuns]{MIfuns} > \link[=helpPrepare]{Data}

I want to ....
\itemize{
\item \link[=ibw]{calculate} various clinical measurements, e.g., body size
\item \link[=css]{calculate} one-compartment model properties
\item \link[=check.subjects]{summarize} a table of subject data
\item \link[=Tmax]{calculate} Tmax
\item \link[=Tmin]{calculate} Tmin
\item \link[=AUC]{calculate} AUC
}
}
