\name{helpPrepare}
\alias{helpPrepare}
\docType{package}
\title{MIfuns: Data Preparation}
\description{Tools for preparing pharmacometric analysis data sets.}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item work with \link[=helpVector]{vectors}    
\item work with \link[=helpDataFrame]{data.frames}
\item work with \link[=helpMatrix]{matrices}
\item \link[=helpClasses]{use} classes specialized for data assembly
\item \link[=helpPharmacometric]{investigate} pharmacometric properties
\item \link[=helpStrategic]{perform} strategic data manipulations
}
}
