\name{helpReport}
\alias{helpReport}
\docType{package}
\title{MIfuns: Reporting}
\description{Tools for summarizing and reporting pharmacometric analyses.}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=params]{identify} documented parameter names
\item \link[=lookup]{lookup} some other form of parameter names
\item \link[=rlog]{generate} a combined run log from multiple NONMEM runs
\item \link[=partab]{create} a model parameter table
\item \link[=wiki2latex]{convert} wikimath to latex or plotmath
\item \link[=justUnits]{manipulate} wikimath elements
\item \link[=iterations]{list} the iteration statistics in the NONMEM output
\item \link[=omegacor]{convert} the NONMEM output omega or sigma covariance matrix to correlations
\item \link[=ftable2data.frame]{reformat} an ftable for printing with latex
\item \link[=ltable.data.frame]{convert} a data.frame for printing with latex
}
}
