\name{helpVector}
\alias{helpVector}
\docType{package}
\title{MIfuns: Working with Vectors}
\description{Tools for processing vectors.}
\details{
\link[=MIfuns]{MIfuns} > \link[=helpPrepare]{Data}

I want to ....
\itemize{
\item \link[=text2decimal]{convert} mixed text to decimal
\item \link[=contains]{test} a character vector for occurences of a pattern
\item \link[=forbak]{impute} missing vector values
\item \link[=runhead]{check} some properties of vectors
\item \link[=miTemporal]{create} date, time, and date-time objects
\item \link[=map]{map} one set of values to another
}
}
