\name{helpVisualize}
\alias{helpVisualize}
\docType{package}
\title{MIfuns: Plotting}
\description{Tools for creating plots.}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=PLOTR]{create} diagnostic plots for finished model runs
\item \link[=covplot]{make} \sQuote{forest plots} of covariate effects
\item \link[=panel.stratify]{handle} each level of a stripplot independently 
\item \link[=plot.nm]{plot} nm objects
\item \link[=xyplot.ext]{plot} the parameter search history for a NONMEM7 run
}
}
