\name{ltable}
\alias{ltable}
\title{Create a Latex Table}
\description{Create a latex table.}
\usage{
ltable(x, ...)
}
\arguments{
  \item{x}{dispatch argument}
  \item{\dots}{ignored}
}
\details{\code{ltable} is generic.}
\value{character}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{	\itemize{
			\item \code{\link{ltable.data.frame}}
	}
}
\keyword{manip}

