\name{MIfuns-package}
\alias{MIfuns-package}
\alias{MIfuns}
\docType{package}
\title{Pharmacometric Tools for Data Preparation, Modeling, Simulation, and Reporting}
\description{
This package includes pharmacometric tools for common data preparation tasks, stratified
bootstrap resampling of data sets, NONMEM control stream
creation/editing, NONMEM model execution, creation of standard and
user-defined diagnostic plots, execution and summary of bootstrap and
predictive check results, implementation of simulations from posterior
parameter distributions, reporting of output tables, and creation of
detailed analysis logs.

Deprecated.  Please see the replacement package \pkg{metrumrg}.
}
\details{

I want to ....
\itemize{
\item \link[=helpPrepare]{prepare} or manipulate data
\item \link[=helpModel]{model} data in NONMEM
\item \link[=helpSimulate]{simulate} from existing models
\item \link[=helpVisualize]{plot} model inputs or outputs
\item \link[=helpQuantify]{calculate} general statistics
\item \link[=helpReport]{report} modeling and simulation results systematically
\item \link[=helpAdminister]{administer} related files and file systems
\item \link[=helpList]{list} all functions
}  
\tabular{ll}{
Package: \tab MIfuns\cr
Type: \tab Package\cr
Version: \tab 5.1\cr
Date: \tab 2011-09-07\cr
License: \tab GPL\cr
}
}
  
  \author{Developed by Metrum Institute (Bill Knebel, Tim Bergsma, Leonid Gibianski, others). Maintainer: Tim Bergsma <timb@metrumrg.com>}
  
  \references{http://mifuns.googlecode.com}
  \keyword{manip}
