\name{nix}
\alias{nix}

\title{Identify Unix-like Platforms}
\description{
 Tells whether the platform OS is of type 'unix'.
}
\usage{
nix()
}
\value{logical}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item \code{\link{NONR}}
	}
}
\examples{
nix()
}
\keyword{manip}

