\name{setCwres}
\alias{setCwres}

\title{Append Conditional Weighted Residuals to an Appropriate File}
\description{
 Appends conditional weighted residuals to an appropriate file.
}
\usage{
setCwres(cwres, file)
}

\arguments{
  \item{cwres}{vector of conditional weighted residuals}
  \item{file}{file to append}
}
\value{Used for side effects}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item \code{\link{PLOTR}}
	}
}
\keyword{manip}

