\name{spaces}
\alias{spaces}
\title{Generate a String of Spaces}
\description{Generate a string of spaces.}
\usage{
spaces(x)
}
\arguments{
  \item{x}{scalar numeric}
}
\details{generates a string of \code{x} spaces}
\value{character}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{	\itemize{
			\item \code{\link{command}}
			\item \code{\link{wrap}}
	}
}
\examples{
spaces(3)
}
\keyword{manip}

