\name{strain}
\alias{strain}

\title{Reduce x to Those Elements that Occur in Options}
\description{
 a helper function that filters x on options
}
\usage{
strain(x, options)
}

\arguments{
  \item{x}{character}
  \item{options}{character}
}
\value{character}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item \code{\link{PLOTR}}
	}
}
\keyword{manip}

