\name{tabular}
\alias{tabular}
\title{Create a Tabular Environment}
\description{Create a tabular environment.}
\usage{
tabular(x, ...)
}
\arguments{
  \item{x}{dispatch argument}
  \item{\dots}{ignored}
}
\details{Generic}
\value{character}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{	\itemize{
			\item \code{\link{tabular.data.frame}}
	}
}
\keyword{manip}

