\name{win}
\alias{win}

\title{Identify Windows Platforms}
\description{
	Tells whether the platform OS is of type \code{windows}
}
\usage{
win()
}
\value{logical}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item \code{\link{NONR}}
	}
}
\examples{
win()
}
\keyword{manip}

