\name{getdname}
\alias{getdname}

\title{Extract the Path and File from the DATA Block in the Control Stream Associated with a Filename}
\description{
	Extracts the data path from the control stream, e.g., for use with \code{getCovs}.
}
\usage{
getdname(filename)
}
\arguments{
  \item{filename}{path for the control stream}
}
\value{character}
\references{\url{http://mifuns.googlecode.com}}
\author{Tim Bergsma}
\seealso{
	\itemize{
		\item \code{\link{PLOTR}}
	}
}
\keyword{manip}

