\name{helpDataFrame}
\alias{helpDataFrame}
\docType{package}
\title{MIfuns: Working with Data Frames}
\description{Tools for processing data.frames.}
\details{
\link[=MIfuns]{MIfuns} > \link[=helpPrepare]{Data}

I want to ....
\itemize{
\item \link[=shuffle]{move} columns in a data frame
\item \link[=colname<-]{rename} selected columns
\item \link[=fixedwidth]{format} a data.frame as fixed-width
\item \link[=aug]{add} named elements to a data.frame or list
}
}
