\name{helpMatrix}
\alias{helpMatrix}
\docType{package}
\title{MIfuns: Working with Matrices}
\description{Tools for processing matrices.}
\details{
\link[=MIfuns]{MIfuns} > \link[=helpPrepare]{Data}

I want to ....
\itemize{
\item \link[=half.matrix]{interconvert} between square matrices and triangular matrices 
}
}
