\name{helpModel}
\alias{helpModel}
\docType{package}
\title{MIfuns: Modeling}
\description{Tools for preparing, executing, and analyzing models.}

\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=helpPrepare]{assemble} an analysis data set
\item \link[=metaSub.character]{make} multiple, systematic substitutions in text files, e.g., control streams
\item \link[=as.nmctl]{convert} NONMEM control streams to modifiable objects
\item \link[=NONR]{run} NONMEM and create diagnostic plots
\item \link[=runstate]{determine} the status of a run
\item \link[=compute.cwres]{compute} conditional weighted residuals in NONMEM 6, etc.
\item \link[=helpVisualize]{visualize} model results
}
}

