\name{helpQuantify}
\alias{helpQuantify}
\docType{package}
\title{MIfuns: Statistics}
\description{Tools for calculating statistics.}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=bin]{assign} values in a vector to quantile bins
\item \link[=snap]{coerce} values to nearest of candidates
}
}
