\name{helpSimulate}
\alias{helpSimulate}
\docType{package}
\title{MIfuns: Simulation}
\description{Tools for conducting simulations.}
\details{
\link[=MIfuns]{MIfuns}

I want to ....
\itemize{
\item \link[=resample.data.frame]{create} replicate datasets by stratified sampling with replacement    
\item \link[=as.matrix.halfmatrix]{convert} a vectorized triangular matrix of random effects to a full matrix
\item \link[=simpar]{create} parameters for simulation with uncertainity
\item \link[=metaSub.character]{make} multiple, systematic substitutions in text files, e.g., control streams
\item \link[=NONR]{run} NONMEM and create diagnostic plots
\item \link[=helpVisualize]{visualize} simulation results
}
}
